pull de service listas

+++ src/services/ClienteServices.java
@@ -2,9 +2,16 @@ package services;
 
 import Controllers.Validation;
 import application.clases.Cliente;
+import application.clases.Propietario;
+import application.clases.TipoDNI;
+import application.clases.TipoInmueble;
 import application.dao.ClienteDAO;
+import dto.ClienteDTO;
+import dto.PropietarioDTO;
+
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class ClienteServices {
 	private static ClienteServices instance;
@@ -17,22 +24,57 @@ public class ClienteServices {
 		}
 		return instance;
 	}
-	public void createCliente(Cliente cliente) {
-		 clientedao.createCliente(cliente);
+	public int createCliente(ClienteDTO entrada) {
+		TipoDNI tipoDNI= TipoDNI.valueOf(entrada.getTipoDNI());
+		if(chequearDuplicado( tipoDNI,entrada.getDni())) {return -1;}//si existe un duplicado retorna -1
+		else {
+			Cliente cliente = toCliente(entrada);
+			 clientedao.createCliente(cliente);
+			return 1;}
+		
 	};
-	public void deleteCliente(Cliente cliente) {
-		 clientedao.deleteCliente(cliente);
+	public int deleteCliente(int id) {
+		Cliente cliente = clientedao.getClienteById(id);
+		if(cliente !=null) {
+			clientedao.deleteCliente(cliente);
+			return 1;
+		}else {return -1;}
 	};
-	public void updateCliente(Cliente cliente) {
-		clientedao.updateCliente(cliente);
+	public int updateCliente(ClienteDTO entrada) {
+		Cliente og = clientedao.getClienteById(entrada.getId());
+		if(og !=null) {
+			Cliente cliente = toCliente(entrada);
+			cliente.setDni(og.getDni());
+			cliente.setTipodni(og.getTipodni());
+			clientedao.updateCliente(cliente);
+			return 1;
+		}else {return -1;}
+		
 	};
-	public List<Cliente> getClientes(String nombre, String apellido) {
-		ArrayList<String> criterios = new ArrayList<String>();
-		criterios.add(nombre);
-		criterios.add(apellido);
-		return clientedao.getClientes(criterios);
+	public List<ClienteDTO> listClientes(){
+		return clientedao.getAllClientes().stream()
+	            .map(cliente -> new ClienteDTO(cliente))
+	            .collect(Collectors.toList());
 	}
-	public Cliente getClienteById(int id) {
-		return clientedao.getClienteById(id);
+	public List<ClienteDTO> getClientes(String tipoDNI,String dni,String nombre, String apellido) {
+		TipoDNI tipo= TipoDNI.valueOf(tipoDNI.replace(" ", "_"));
+
+		return clientedao.getCliente(dni,tipo,nombre,apellido).stream()
+	            .map(cliente -> new ClienteDTO(cliente))
+	            .collect(Collectors.toList());
 	}
+	public ClienteDTO getClienteById(int id) {
+		return new ClienteDTO(clientedao.getClienteById(id));
+	}
+	private Cliente toCliente( ClienteDTO entrada) {
+		TipoInmueble tipoInmueble= TipoInmueble.valueOf(entrada.getTipoInmuebleBuscado());
+		TipoDNI tipoDNI= TipoDNI.valueOf(entrada.getTipoDNI());
+		return new Cliente( entrada.getNombre(), entrada.getApellido(), entrada.getDni(),tipoDNI,entrada.getTelefono(),entrada.getEmail(),
+				entrada.getMontoDisponible(),tipoInmueble, entrada.getLocalidadBuscada(), entrada.getBarrios(),entrada.getCaracteristicasDeseadas());
+	}
+	private boolean chequearDuplicado( TipoDNI tipo,String dni) {
+		List<Cliente> lista=clientedao.getCliente(dni,tipo,null,null);
+		if(lista==null){return false;}
+		else { return true;}
+		}
 }
